Handlers:
handleChange:

Triggered when the value of an input field changes.
Validates that the entered value is either an alphanumeric character or empty.
Updates the OTP state with the new value.
Automatically moves focus to the next input field if a digit is entered, except for the last input.
handleKeyDown:

Triggered on keydown events for input fields.
Specifically handles the "Backspace" key to move focus to the previous input field if the current field is empty and the user is deleting a character.
submitHandler:

Triggered when the form is submitted.
Prevents the default form submission behavior.
Joins the OTP array into a single string and attempts to verify the email using the verifyEmail function from useUserStore.
Navigates to the home page upon successful verification. Logs errors to the console if verification fails.