Event Handlers:
fileChangeHandler: Handles file input changes. It reads the selected file and updates the profileData and selectedProfilePicture states with the new image data URL.
changeHandler: Updates the profileData state with the changes made to the text input fields.
updateProfileHandler: Handles the form submission. It prevents the default form submission, sets the loading state to true, calls the updateProfile function from the useUserStore to update the profile, and then sets the loading state to false. It also catches and handles any errors.