Components and Hooks useParams Hook:
Extracts parameters from the URL. Here, it’s used to get the search text (params.text) that was passed in the route. This parameter could represent a search term or category for filtering restaurants.
State Management:

searchQuery: Holds the current search query entered by the user in the input field.
loading: Boolean state indicating whether the data is being fetched or not.
searchedRestaurant: Contains the results of the search query, including the data and any associated metadata.
searchRestaurant: A function from the useRestaurantStore hook that performs the search operation based on the current search query and applied filters.
setAppliedFilter: Function to set the filters that have been applied to the search results.
appliedFilter: Array of filters that have been applied to the search results.