Imports useRestaurantStore:
A custom hook that provides access to the global state related to restaurants. It allows the component to read and update the list of applied filters and reset them.
Button:

A UI component used for the "Reset" button. It’s styled and customized to handle user interactions.
Checkbox:

A UI component representing checkboxes for filtering options. It allows users to select or deselect filter options.
Label:

A UI component used to provide descriptive text next to each checkbox. It’s used to display the name of the filter options.