Component Structure Props:
menus:
An array of MenuItem objects passed as a prop. Each MenuItem represents a menu option available at the restaurant.
State and Handlers:

addToCart:
A function from useCartStore that adds a menu item to the shopping cart. It is invoked when a user clicks the "Add to Cart" button.
navigate:
A function from useNavigate that programmatically navigates the user to the /cart route when an item is added to the cart.